Split storage class to smaller classes
