+++ data/tasks.txt
@@ -2,3 +2,5 @@ T | 0 | haha
 T | 0 | me /to lol
 E | 0 | frrom | 10-10-2005 12:30 | 10-11-2005 12:30
 T | 0 | lol
+T | 0 | lol
+T | 0 | haha
