+++ src/main/java/darkpool/command/ListCommand.java
@@ -1,7 +1,7 @@
 package darkpool.command;
 
 import darkpool.gui.Gui;
-import darkpool.util.Storage;
+import darkpool.storage.Storage;
 import darkpool.tasklist.TaskList;
 
 public class ListCommand extends Command {
