+++ src/main/java/darkpool/parser/ExitParser.java
@@ -2,7 +2,7 @@ package darkpool.parser;
 
 import darkpool.command.Command;
 import darkpool.command.ExitCommand;
-import darkpool.util.DarkpoolException;
+import darkpool.DarkpoolException;
 
 public class ExitParser {
