+++ src/main/java/darkpool/parser/FindParser.java
@@ -2,7 +2,7 @@ package darkpool.parser;
 
 import darkpool.command.Command;
 import darkpool.command.FindCommand;
-import darkpool.util.DarkpoolException;
+import darkpool.DarkpoolException;
 
 public class FindParser {
