+++ src/main/java/darkpool/parser/ListParser.java
@@ -2,7 +2,7 @@ package darkpool.parser;
 
 import darkpool.command.Command;
 import darkpool.command.ListCommand;
-import darkpool.util.DarkpoolException;
+import darkpool.DarkpoolException;
 
 public class ListParser {
