+++ src/main/java/darkpool/parser/TodoParser.java
@@ -2,7 +2,7 @@ package darkpool.parser;
 
 import darkpool.command.AddCommand;
 import darkpool.command.Command;
-import darkpool.util.DarkpoolException;
+import darkpool.DarkpoolException;
 
 
 public class TodoParser {
