+++ src/main/java/darkpool/task/Event.java
@@ -3,7 +3,7 @@ package darkpool.task;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 
-import darkpool.util.DarkpoolException;
+import darkpool.DarkpoolException;
 
 /**
  * Represents an event task with a start and end time.
