+++ src/test/java/darkpool/task/EventTest.java
@@ -11,7 +11,7 @@ import java.time.format.DateTimeFormatter;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import darkpool.util.DarkpoolException;
+import darkpool.DarkpoolException;
 
 
 class EventTest {
